thêm API

+++ AcountService/dto/request/discount/DiscountResquest.cs
@@ -4,6 +4,6 @@
     {
         public string code { get; set; }
 
-        public double Percent { get; set; }
+        public decimal Percent { get; set; }
     }
 }
