+++ AcountService/dto/response/Discount/DiscountResponse.cs
@@ -4,6 +4,6 @@
     {
         public int  DiscountId { get; set; }
         public string code { get; set; }
-        public double Percent { get; set; }
+        public decimal Percent { get; set; }
     }
 }
