+++ AcountService/entity/Payment.cs
@@ -6,13 +6,14 @@ namespace AcountService.entity
     public class Payment
     {
         [Key]
+
         public int PaymentId { get; set; } // Khóa chính cho Payment
-        public double Amount { get; set; } // Số tiền thanh toán
+        public decimal Amount { get; set; } // Số tiền thanh toán
         public string Method { get; set; } // Phương thức thanh toán (ví dụ: Credit Card, PayPal)
         public DateTime PaymentDate { get; set; } // Ngày thanh toán
         public bool IsSuccessful { get; set; } // Trạng thái thanh toán
 
-        public int OrderId { get; set; } // Khóa ngoại đến đơn hàng
+         public int OrderId { get; set; } // Khóa ngoại đến đơn hàng
         public Order Order { get; set; } // Liên kết đến đơn hàng
     }
 }
